+++ Sources/GoldSunCore/Models/BrowserDestination.swift
@@ -1,5 +1,6 @@
 import Foundation
 
 public enum BrowserDestination {
+    public static let goldSunStartPage = URL(string: "goldsun://home")!
     public static let chromeWebStore = URL(string: "https://chromewebstore.google.com/")!
 }